requirement.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2022.9.24
cffi==1.15.1
channels==3.0.4
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.3
daphne==3.0.2
Deprecated==1.2.13
Django==4.1.3
gevent==22.10.2
greenlet==2.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
packaging==21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
redis==4.3.4
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
websocket-client==1.2.0
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.5.1
